+++ pkg/idp/provider/github/github.go
@@ -11,6 +11,7 @@ import (
 	"github.com/giantswarm/dex-operator/pkg/idp/provider"
 	"github.com/giantswarm/dex-operator/pkg/idp/provider/github/manifest"
 	"github.com/giantswarm/dex-operator/pkg/key"
+	"github.com/giantswarm/dex-operator/pkg/yaml"
 
 	"github.com/bradleyfalzon/ghinstallation/v2"
 	githubconnector "github.com/dexidp/dex/connector/github"
@@ -18,7 +19,6 @@ import (
 	"github.com/go-logr/logr"
 	githubclient "github.com/google/go-github/v50/github"
 	"github.com/skratchdot/open-golang/open"
-	"gopkg.in/yaml.v2"
 )
 
 const (
@@ -32,6 +32,7 @@ const (
 	ClientIDKey           = "client-id"
 	ClientSecretKey       = "client-secret"
 	DefaultHost           = "github.com"
+	TeamNameFieldSlug     = "slug"
 )
 
 type Github struct {
@@ -174,9 +175,10 @@ func (g *Github) CreateOrUpdateApp(config provider.AppConfig, ctx context.Contex
 				Teams: []string{g.Team},
 			},
 		},
-		RedirectURI: config.RedirectURI,
+		RedirectURI:   config.RedirectURI,
+		TeamNameField: TeamNameFieldSlug,
 	}
-	data, err := yaml.Marshal(connectorConfig)
+	data, err := yaml.MarshalWithJsonAnnotations(connectorConfig)
 	if err != nil {
 		return provider.ProviderApp{}, microerror.Mask(err)
 	}